Handle most cases where parts of the paste are outside the background
Since we can't compute the color difference of an outline point if it's
outside the background bounds, we will simply ignore it in the weighted
average. This removes the artifacts that appeared from such cases.

Note that this doesn't solve the problem where all selected sample
points are outside the background, but the algorithm isn't really meant
for handling such cases so this is the best we can do.